added auth gate

+++ routes/web.php
@@ -25,19 +25,19 @@ Route::get('/locale/{locale}', function ($locale) {
 })->name('locale');
 // Auth middleware for authenticated users
 Route::middleware(['auth', 'verified'])->group(function () {
-    Route::get('/roles', [App\Http\Controllers\RoleController::class, 'index'])->name('role');
-    Route::get('/profile', [App\Http\Controllers\UserController::class, 'profile'])->name('profile');
+    Route::get('/roles', [App\Http\Controllers\RoleController::class, 'index'])->name('role')->middleware(['role:admin']);
     Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
     Route::get('/queries', [App\Http\Controllers\QueryController::class, 'index'])->name('query');
 
 
-    Route::get('users', [App\Http\Controllers\UserController::class, 'index'])->name('user.index');
-    Route::get('users/create', [App\Http\Controllers\UserController::class, 'create'])->name('user.create');
-    Route::post('users/store', [App\Http\Controllers\UserController::class, 'store'])->name('user.store');
-    Route::get('users/{id}/edit', [App\Http\Controllers\UserController::class, 'edit'])->name('user.edit');
-    Route::put('users/{id}', [App\Http\Controllers\UserController::class, 'update'])->name('user.update');
+    Route::get('users', [App\Http\Controllers\UserController::class, 'index'])->name('user.index')->middleware(['role:admin']);
+    Route::get('users/create', [App\Http\Controllers\UserController::class, 'create'])->name('user.create')->middleware(['role:admin']);
+    Route::post('users/store', [App\Http\Controllers\UserController::class, 'store'])->name('user.store')->middleware(['role:admin']);
+    Route::get('users/{id}/edit', [App\Http\Controllers\UserController::class, 'edit'])->name('user.edit')->middleware(['role:admin']);
+    Route::put('users/{id}', [App\Http\Controllers\UserController::class, 'update'])->name('user.update')->middleware(['role:admin']);
+    Route::put('users/{id}/change-password', [App\Http\Controllers\UserController::class, 'changeUserPassword'])->name('user.change-pass')->middleware(['role:admin']);
     Route::get('users/{id}/view', [App\Http\Controllers\UserController::class, 'view'])->name('user.view');
-    Route::delete('users/{id}', [App\Http\Controllers\UserController::class, 'destroy'])->name('user.destroy');
+    Route::delete('users/{id}', [App\Http\Controllers\UserController::class, 'destroy'])->name('user.destroy')->middleware(['role:admin']);
 
 });
